+++ kafka-client/src/main/java/dev/responsive/kafka/store/CommitBuffer.java
@@ -18,11 +18,7 @@ package dev.responsive.kafka.store;
 
 import static dev.responsive.kafka.config.ResponsiveConfig.MAX_CONCURRENT_REMOTE_WRITES_CONFIG;
 
-import dev.responsive.db.CassandraClient;
-import dev.responsive.db.KeySpec;
-import dev.responsive.db.MetadataRow;
-import dev.responsive.db.RemoteSchema;
-import dev.responsive.db.WriterFactory;
+import dev.responsive.db.*;
 import dev.responsive.db.partitioning.SubPartitioner;
 import dev.responsive.kafka.clients.SharedClients;
 import dev.responsive.kafka.config.ResponsiveConfig;
