+++ kafka-client/src/test/java/dev/responsive/kafka/store/LwtWriterTest.java
@@ -28,11 +28,8 @@ import com.datastax.oss.driver.api.core.cql.AsyncResultSet;
 import com.datastax.oss.driver.api.core.cql.BoundStatement;
 import com.datastax.oss.driver.api.core.cql.Statement;
 import com.datastax.oss.driver.internal.core.cql.DefaultBatchStatement;
-import dev.responsive.db.BytesKeySpec;
-import dev.responsive.db.CassandraClient;
-import dev.responsive.db.CassandraKeyValueSchema;
-import dev.responsive.db.KeySpec;
-import dev.responsive.db.WriterFactory;
+import dev.responsive.db.*;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
